@@ -7,10 +7,10 @@
 #include <boost/log/trivial.hpp>
 
 #include <cstdint>
+#include <format>
 #include <stdexcept>
 #include <string>
 
-#include <fmt/core.h>
 #include <zcpm/terminal/terminal.hpp>
 
 // The constructor sets up various tables in RAM for the BIOS. This is the layout (from low address to high):
@@ -50,7 +50,7 @@ Bios::Bios(Hardware* p_hardware, terminal::Terminal* p_terminal) : m_phardware(p
     // maybe stack should be elsewhere?
     m_stubs_base = base + 0x0100;
 
-    BOOST_LOG_TRIVIAL(trace) << fmt::format("Rewriting BIOS jump table at {:04X}", base);
+    BOOST_LOG_TRIVIAL(trace) << std::format("Rewriting BIOS jump table at {:04X}", base);
 
     // Write our own jump table over the top of whatever is there currently, making it point to another
     // one ("stubs") higher in memory, and that's the one that is intercepted (which should reduce the
@@ -121,7 +121,7 @@ Bios::Bios(Hardware* p_hardware, terminal::Terminal* p_terminal) : m_phardware(p
         m_phardware->write_byte(i, 0x00);
     }
 
-    BOOST_LOG_TRIVIAL(trace) << fmt::format("BIOS jump table {:04X}..{:04X}, BIOS stubs {:04X}..{:04X}, DPH etc {:04X}..{:04X}",
+    BOOST_LOG_TRIVIAL(trace) << std::format("BIOS jump table {:04X}..{:04X}, BIOS stubs {:04X}..{:04X}, DPH etc {:04X}..{:04X}",
                                             m_discovered_base,
                                             m_discovered_base + table_size * 3 - 1,
                                             m_stubs_base,
@@ -129,7 +129,7 @@ Bios::Bios(Hardware* p_hardware, terminal::Terminal* p_terminal) : m_phardware(p
                                             m_dph_base,
                                             m_dph_top);
 
-    BOOST_LOG_TRIVIAL(trace) << fmt::format("     dirbf={:04X} hdblk={:04X} chkhd1={:04X} allhd1={:04X}", dirbf, hdblk, chkhd1, allhd1);
+    BOOST_LOG_TRIVIAL(trace) << std::format("     dirbf={:04X} hdblk={:04X} chkhd1={:04X} allhd1={:04X}", dirbf, hdblk, chkhd1, allhd1);
 
     // Set up monitoring of the DPH etc stuff
     m_phardware->add_watch_read(m_dph_base, m_dph_top - m_dph_base + 1);
@@ -165,7 +165,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     // The address is within the BIOS stubs. Work out what BIOS function is being called
     const unsigned int fn = address - m_stubs_base;
 
-    const auto prefix(fmt::format("BIOS fn#{:d} ", fn));
+    const auto prefix(std::format("BIOS fn#{:d} ", fn));
     std::string msg;
 
     switch (fn)
@@ -201,7 +201,7 @@ bool Bios::check_and_handle(std::uint16_t address)
         // Block until a character is ready, and then return it in A
         m_phardware->m_processor->reg_a() = m_pterminal->get_char();
         const auto ch = m_phardware->m_processor->get_a();
-        msg = fmt::format("CONIN({:02X})", ch);
+        msg = std::format("CONIN({:02X})", ch);
         log_bios_call(prefix, msg);
     }
     break;
@@ -210,11 +210,11 @@ bool Bios::check_and_handle(std::uint16_t address)
         const auto ch = static_cast<char>(m_phardware->m_processor->get_c());
         if (ch >= ' ')
         {
-            msg = fmt::format("CONOUT({:02X} '{:c}')", ch, ch);
+            msg = std::format("CONOUT({:02X} '{:c}')", ch, ch);
         }
         else
         {
-            msg = fmt::format("CONOUT({:02X})", ch);
+            msg = std::format("CONOUT({:02X})", ch);
         }
         log_bios_call(prefix, msg);
         m_pterminal->print(ch);
@@ -231,7 +231,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     {
         const auto disk = m_phardware->m_processor->get_c(); // Disk index; 0=A, 1=B, etc
         const auto flag = m_phardware->m_processor->get_e(); // Bit 0 is the "has been logged in before" flag
-        msg = fmt::format("SELDSK(disk={:02X},flag={:02X})", disk, flag);
+        msg = std::format("SELDSK(disk={:02X},flag={:02X})", disk, flag);
         log_bios_call(prefix, msg);
         fn_seldsk(disk, flag);
     }
@@ -239,7 +239,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     case 10:
     {
         const auto bc = m_phardware->m_processor->get_bc();
-        msg = fmt::format("SETTRK({:04X})", bc);
+        msg = std::format("SETTRK({:04X})", bc);
         log_bios_call(prefix, msg);
         fn_settrk(bc);
     }
@@ -247,7 +247,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     case 11:
     {
         const auto bc = m_phardware->m_processor->get_bc();
-        msg = fmt::format("SETSEC({:04X})", bc);
+        msg = std::format("SETSEC({:04X})", bc);
         log_bios_call(prefix, msg);
         fn_setsec(bc);
     }
@@ -255,7 +255,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     case 12:
     {
         const auto bc = m_phardware->m_processor->get_bc();
-        msg = fmt::format("SETDMA({:04X})", bc);
+        msg = std::format("SETDMA({:04X})", bc);
         log_bios_call(prefix, msg);
         fn_setdma(bc);
     }
@@ -270,7 +270,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     case 14:
     {
         const auto c = m_phardware->m_processor->get_c();
-        msg = fmt::format("WRITE({:02X})", c);
+        msg = std::format("WRITE({:02X})", c);
         log_bios_call(prefix, msg);
         m_phardware->m_processor->reg_a() = fn_write(c);
     }
@@ -279,7 +279,7 @@ bool Bios::check_and_handle(std::uint16_t address)
     {
         const auto bc = m_phardware->m_processor->get_bc();
         const auto de = m_phardware->m_processor->get_de();
-        msg = fmt::format("SECTRAN({:04X},{:04X})", bc, de);
+        msg = std::format("SECTRAN({:04X},{:04X})", bc, de);
         log_bios_call(prefix, msg);
         const auto physical_sector_number = fn_sectran(bc, de);
         m_phardware->m_processor->reg_hl() = physical_sector_number;
@@ -361,7 +361,7 @@ std::uint8_t Bios::fn_read()
     // See http://www.seasip.info/Cpm/bios.html#read
     // Read the sector at m_track/m_sector into m_dma, return success status (which ends up in register A)
 
-    BOOST_LOG_TRIVIAL(trace) << fmt::format("Read TRACK:{:04X},SECTOR:{:04X} into {:04X}", m_track, m_sector, m_dma);
+    BOOST_LOG_TRIVIAL(trace) << std::format("Read TRACK:{:04X},SECTOR:{:04X} into {:04X}", m_track, m_sector, m_dma);
 
     // Allocate a sector-size chunk of memory, into which data is read
     Disk::SectorData buffer{};
@@ -378,7 +378,7 @@ std::uint8_t Bios::fn_write(std::uint8_t /*deblocking*/)
     // See http://www.seasip.info/Cpm/bios.html#write
     // Write from m_dma to the sector at m_track/m_sector, return success status (which ends up in register A)
 
-    BOOST_LOG_TRIVIAL(trace) << fmt::format("Write TRACK:{:04X},SECTOR:{:04X} from {:04X}", m_track, m_sector, m_dma);
+    BOOST_LOG_TRIVIAL(trace) << std::format("Write TRACK:{:04X},SECTOR:{:04X} from {:04X}", m_track, m_sector, m_dma);
 
     // To help with debugging
     m_phardware->dump(m_dma, Disk::SectorSize);
