@@ -4,11 +4,11 @@
 
 #include <array>
 #include <cstdint>
+#include <format>
 #include <ostream>
 #include <string>
 #include <vector>
 
-#include <fmt/core.h>
 #include <zcpm/core/processor.hpp>
 #include <zcpm/core/registers.hpp>
 
@@ -28,13 +28,13 @@ const std::map<std::uint8_t, const char*> DdFdCbLogicals = {
 // A 8-bit literal as a 2 digit hex string
 auto byte(std::uint8_t x)
 {
-    return fmt::format("{:02X}", x);
+    return std::format("{:02X}", x);
 }
 
 // A 16-bit literal as a 4 digit hex string
 auto word(std::uint8_t low, std::uint8_t high)
 {
-    return fmt::format("{:04X}", high << 8 | low);
+    return std::format("{:04X}", high << 8 | low);
 }
 
 std::string byte_array_to_string(const std::vector<std::uint8_t>& bytes)
@@ -54,30 +54,30 @@ std::string byte_array_to_string(const std::vector<std::uint8_t>& bytes)
 // Dereference NN from (string)
 auto nn_string(std::uint8_t low, std::uint8_t high, const std::string& s)
 {
-    return fmt::format("({:04X}),{}", high << 8 | low, s);
+    return std::format("({:04X}),{}", high << 8 | low, s);
 }
 
 // Dereference (string) from NN
 auto string_nn(std::uint8_t low, std::uint8_t high, const std::string& s)
 {
-    return fmt::format("{},({:04X})", s, high << 8 | low);
+    return std::format("{},({:04X})", s, high << 8 | low);
 }
 
 // Dereference N from (string)
 auto n_string(std::uint8_t n, const std::string& s)
 {
-    return fmt::format("({:02X}),{}", n, s);
+    return std::format("({:02X}),{}", n, s);
 }
 
 // Dereference (string) from N
 // auto string_n(std::uint8_t n, const std::string& s)
 //{
-//  return fmt::format("{},({:02X})", s, n);
+//  return std::format("{},({:02X})", s, n);
 //}
 
 auto hl_ss(std::uint8_t ss)
 {
-    return fmt::format("HL,{}", WordRegMask[ss]);
+    return std::format("HL,{}", WordRegMask[ss]);
 }
 
 auto byte_register(std::uint8_t r)
@@ -93,41 +93,41 @@ auto qq_word_register(std::uint8_t qq)
 // r,n where r is a byte register and n is a 8-bit literal
 auto r_n(std::uint8_t r, std::uint8_t n)
 {
-    return fmt::format("{},{:02X}", ByteRegMask[r], n);
+    return std::format("{},{:02X}", ByteRegMask[r], n);
 }
 
 // r,r where r1 and r2 are both byte registers
 auto r_r(std::uint8_t r1, std::uint8_t r2)
 {
-    return fmt::format("{},{}", ByteRegMask[r1], ByteRegMask[r2]);
+    return std::format("{},{}", ByteRegMask[r1], ByteRegMask[r2]);
 }
 
 // dd,nn where dd is a word register and nn is a 16-bit literal
 auto dd_nn(std::uint8_t dd, std::uint8_t nn_low, std::uint8_t nn_high)
 {
     const std::uint16_t nn = (nn_high << 8) | nn_low;
-    return fmt::format("{},{:04X}", WordRegMask[dd], nn);
+    return std::format("{},{:04X}", WordRegMask[dd], nn);
 }
 
 // (nn),dd where dd is a word register and nn is a 16-bit literal
 auto inn_dd(std::uint8_t dd, std::uint8_t nn_low, std::uint8_t nn_high)
 {
     const std::uint16_t nn = (nn_high << 8) | nn_low;
-    return fmt::format("({:04X}),{}", nn, WordRegMask[dd]);
+    return std::format("({:04X}),{}", nn, WordRegMask[dd]);
 }
 
 // dd,(nn) where dd is a word register and nn is a 16-bit literal
 auto dd_inn(std::uint8_t dd, std::uint8_t nn_low, std::uint8_t nn_high)
 {
     const std::uint16_t nn = (nn_high << 8) | nn_low;
-    return fmt::format("{},({:04X})", WordRegMask[dd], nn);
+    return std::format("{},({:04X})", WordRegMask[dd], nn);
 }
 
 // cc,pq where cc is a 3-bit condition and pq is a 16-bit literal
 auto cc_pq(std::uint8_t cc, std::uint8_t pq_low, std::uint8_t pq_high)
 {
     const std::uint16_t pq = (pq_high << 8) | pq_low;
-    return fmt::format("{},{:04X}", CondMask[cc], pq);
+    return std::format("{},{:04X}", CondMask[cc], pq);
 }
 
 // Relative target as a 4 digit hex value (PC+e)
@@ -135,7 +135,7 @@ auto offset(std::uint16_t pc, std::uint8_t e)
 {
     const int8_t ee = e;
     const std::uint16_t dest = pc + 2 + ee;
-    return fmt::format("{:04X}", dest);
+    return std::format("{:04X}", dest);
 }
 
 // cc,e where cc is a 2-bit condition and e is a relative jump, which
@@ -144,21 +144,21 @@ auto cc_offset(std::uint8_t cc, std::uint8_t e, std::uint16_t pc)
 {
     const int8_t ee = e;
     const std::uint16_t dest = pc + 2 + ee;
-    return fmt::format("{},{:04X}", CondMask[cc], dest);
+    return std::format("{},{:04X}", CondMask[cc], dest);
 }
 
 // "r,(reg+d)"  where r is a 8-bit register index and reg is an index register name and d is an offset
 auto r_ind_offset(std::uint8_t r, const std::string& reg, std::uint8_t offset)
 {
     const int8_t o = offset;
-    return fmt::format("{},({}+{:02X})", ByteRegMask[r], reg, o);
+    return std::format("{},({}+{:02X})", ByteRegMask[r], reg, o);
 }
 
 // "(reg+d),r"  where r is a 8-bit register index and reg is an index register name and d is an offset
 auto ind_offset_r(std::uint8_t r, const std::string& reg, std::uint8_t offset)
 {
     const int8_t o = offset;
-    return fmt::format("({}+{:02X}),{}", reg, o, ByteRegMask[r]);
+    return std::format("({}+{:02X}),{}", reg, o, ByteRegMask[r]);
 }
 
 // TODO: Collate string constants
@@ -179,24 +179,24 @@ std::tuple<size_t, std::string, std::string> disassemble_cb(std::uint8_t op2, st
     {
         const std::uint16_t b = (op2 >> 3) & 0x07;
         const std::uint8_t r = op2 & 0x07;
-        return { 2, "BIT", fmt::format("{:d},", b) + ByteRegMask[r] };
+        return { 2, "BIT", std::format("{:d},", b) + ByteRegMask[r] };
     }
     if ((op2 & 0xC0) == 0x80)
     {
         const std::uint16_t b = (op2 >> 3) & 0x07;
         const std::uint8_t r = op2 & 0x07;
-        return { 2, "RES", fmt::format("{:d},", b) + ByteRegMask[r] };
+        return { 2, "RES", std::format("{:d},", b) + ByteRegMask[r] };
     }
     if ((op2 & 0xC0) == 0xC0)
     {
         const std::uint16_t b = (op2 >> 3) & 0x07;
         const std::uint8_t r = op2 & 0x07;
-        return { 2, "SET", fmt::format("{:d},", b) + ByteRegMask[r] };
+        return { 2, "SET", std::format("{:d},", b) + ByteRegMask[r] };
     }
     if ((op2 & 0xC7) == 0x46)
     {
         const std::uint16_t b = (op2 >> 3) & 0x07;
-        return { 2, "BIT", fmt::format("{:d},(HL)", b) };
+        return { 2, "BIT", std::format("{:d},(HL)", b) };
     }
     if ((op2 & 0xF8) == 0x00)
     {
@@ -240,7 +240,7 @@ std::tuple<size_t, std::string, std::string> disassemble_cb(std::uint8_t op2, st
     }
 
     // No match
-    return { 0, fmt::format("?? CB {:02X}", op2), "" };
+    return { 0, std::format("?? CB {:02X}", op2), "" };
 }
 
 std::tuple<size_t, std::string, std::string> disassemble_ddfd(const std::string& xy,
@@ -254,58 +254,58 @@ std::tuple<size_t, std::string, std::string> disassemble_ddfd(const std::string&
     {
     case 0x09: return { 2, "ADD", xy + ",BC" };
     case 0x19: return { 2, "ADD", xy + ",DE" };
-    case 0x21: return { 4, "LD", fmt::format("{},{:04X}", xy, (op4 << 8) | op3) };
-    case 0x22: return { 4, "LD", fmt::format("({:04X}),{}", (op4 << 8) | op3, xy) };
+    case 0x21: return { 4, "LD", std::format("{},{:04X}", xy, (op4 << 8) | op3) };
+    case 0x22: return { 4, "LD", std::format("({:04X}),{}", (op4 << 8) | op3, xy) };
     case 0x23: return { 2, "INC", xy };
     case 0x24: return { 2, "INC", xy + "H" };
     case 0x25: return { 2, "DEC", xy + "H" };
-    case 0x26: return { 3, "LD", fmt::format("{}H,{:02X}", xy, op3) };
+    case 0x26: return { 3, "LD", std::format("{}H,{:02X}", xy, op3) };
     case 0x29: return { 2, "ADD", xy + "," + xy };
-    case 0x2A: return { 4, "LD", fmt::format("{},({:04X})", xy, (op4 << 8) | op3) };
+    case 0x2A: return { 4, "LD", std::format("{},({:04X})", xy, (op4 << 8) | op3) };
     case 0x2B: return { 2, "DEC", xy };
     case 0x2C: return { 2, "INC", xy + "L" };
     case 0x2D: return { 2, "DEC", xy + "L" };
-    case 0x2E: return { 3, "LD", fmt::format("{}L,{:02X}", xy, op3) };
-    case 0x34: return { 3, "INC", fmt::format("({}+{:02X})", xy, op3) };
-    case 0x35: return { 3, "DEC", fmt::format("({}+{:02X})", xy, op3) };
-    case 0x36: return { 4, "LD", fmt::format("({}+{:02X}),{:02X}", xy, op3, op4) };
+    case 0x2E: return { 3, "LD", std::format("{}L,{:02X}", xy, op3) };
+    case 0x34: return { 3, "INC", std::format("({}+{:02X})", xy, op3) };
+    case 0x35: return { 3, "DEC", std::format("({}+{:02X})", xy, op3) };
+    case 0x36: return { 4, "LD", std::format("({}+{:02X}),{:02X}", xy, op3, op4) };
     case 0x39: return { 2, "ADD", xy + ",SP" };
-    case 0x86: return { 3, "ADD", fmt::format("A,({}+{:02X})", xy, op3) };
-    case 0x8E: return { 3, "ADC", fmt::format("A,({}+{:02X})", xy, op3) };
+    case 0x86: return { 3, "ADD", std::format("A,({}+{:02X})", xy, op3) };
+    case 0x8E: return { 3, "ADC", std::format("A,({}+{:02X})", xy, op3) };
     case 0x94: return { 2, "SUB", xy + "H" };
     case 0x95: return { 2, "SUB", xy + "L" };
-    case 0x96: return { 3, "SUB", fmt::format("A,({}+{:02X})", xy, op3) };
-    case 0x9E: return { 3, "SBC", fmt::format("A,({}+{:02X})", xy, op3) };
-    case 0xA6: return { 3, "AND", fmt::format("A,({}+{:02X})", xy, op3) };
-    case 0xAE: return { 3, "XOR", fmt::format("A,({}+{:02X})", xy, op3) };
-    case 0xB6: return { 3, "OR", fmt::format("A,({}+{:02X})", xy, op3) };
-    case 0xBE: return { 3, "CP", fmt::format("A,({}+{:02X})", xy, op3) };
+    case 0x96: return { 3, "SUB", std::format("A,({}+{:02X})", xy, op3) };
+    case 0x9E: return { 3, "SBC", std::format("A,({}+{:02X})", xy, op3) };
+    case 0xA6: return { 3, "AND", std::format("A,({}+{:02X})", xy, op3) };
+    case 0xAE: return { 3, "XOR", std::format("A,({}+{:02X})", xy, op3) };
+    case 0xB6: return { 3, "OR", std::format("A,({}+{:02X})", xy, op3) };
+    case 0xBE: return { 3, "CP", std::format("A,({}+{:02X})", xy, op3) };
     case 0xCB:
     {
         // TODO: move to a new method of its own? We'll see how well this scales...
         if (DdFdCbLogicals.contains(op4))
         {
-            return { 4, DdFdCbLogicals.at(op4), fmt::format("({}+{:02X})", xy, op3) };
+            return { 4, DdFdCbLogicals.at(op4), std::format("({}+{:02X})", xy, op3) };
         }
         else if ((op4 & 0xC0) == 0x80)
         {
             const auto b = (op4 >> 3) & 0x07;
-            return { 4, "RES", fmt::format("{:d},({}+{:02X})", b, xy, op3) };
+            return { 4, "RES", std::format("{:d},({}+{:02X})", b, xy, op3) };
         }
         else if ((op4 & 0xC0) == 0x40)
         {
             const auto b = (op4 >> 3) & 0x07;
-            return { 4, "BIT", fmt::format("{:d},({}+{:02X})", b, xy, op3) };
+            return { 4, "BIT", std::format("{:d},({}+{:02X})", b, xy, op3) };
         }
         else if ((op4 & 0xC0) == 0xC0)
         {
             const auto b = (op4 >> 3) & 0x07;
-            return { 4, "SET", fmt::format("{:d},({}+{:02X})", b, xy, op3) };
+            return { 4, "SET", std::format("{:d},({}+{:02X})", b, xy, op3) };
         }
         else
         {
             // Unimplemented sequence
-            const auto message = fmt::format("Unimplemented {:02X} {:02X} {:02X} {:02X}", op1, op2, op3, op4);
+            const auto message = std::format("Unimplemented {:02X} {:02X} {:02X} {:02X}", op1, op2, op3, op4);
             throw std::logic_error(message);
         }
     }
@@ -325,12 +325,12 @@ std::tuple<size_t, std::string, std::string> disassemble_ddfd(const std::string&
         const auto table = (op1 == 0xDD) ? DDByteRegMask : FDByteRegMask;
         if (p1 == 0x06)
         {
-            const auto lhs = fmt::format("({}+{:02X})", xy, op3);
+            const auto lhs = std::format("({}+{:02X})", xy, op3);
             return { 3, "LD", lhs + "," + std::string(ByteRegMask[p2]) };
         }
         else if (p2 == 0x06)
         {
-            const auto rhs = fmt::format("({}+{:02X})", xy, op3);
+            const auto rhs = std::format("({}+{:02X})", xy, op3);
             return { 3, "LD", std::string(ByteRegMask[p1]) + "," + rhs };
         }
         else
@@ -341,7 +341,7 @@ std::tuple<size_t, std::string, std::string> disassemble_ddfd(const std::string&
     if ((op4 & 0xC0) == 0x40)
     {
         const auto b = (op4 >> 3) & 0x07;
-        return { 4, "BIT", fmt::format("{:X},({}+{:02X})", b, xy, op3) };
+        return { 4, "BIT", std::format("{:X},({}+{:02X})", b, xy, op3) };
     }
     if ((op2 & 0xC7) == 0x46)
     {
@@ -355,7 +355,7 @@ std::tuple<size_t, std::string, std::string> disassemble_ddfd(const std::string&
     }
 
     // Unimplemented sequence
-    const auto message = fmt::format("Unimplemented {:02X} {:02X} {:02X} {:02X}", op1, op2, op3, op4);
+    const auto message = std::format("Unimplemented {:02X} {:02X} {:02X} {:02X}", op1, op2, op3, op4);
     throw std::logic_error(message);
 }
 
@@ -403,7 +403,7 @@ std::tuple<size_t, std::string, std::string> disassemble_ed(std::uint8_t op2, st
     }
 
     // No match
-    return { 0, fmt::format("?? ED {:02X}", op2), "" };
+    return { 0, std::format("?? ED {:02X}", op2), "" };
 }
 
 // Given 4 bytes at the current PC (plus the PC), returns two human-readable "words" of disassembly,
@@ -595,7 +595,7 @@ std::tuple<size_t, std::string, std::string> disassemble(std::uint8_t op1,
     }
 
     // Unhandled instruction
-    return { 0, fmt::format("?TODO({:02X},{:02X},{:02X})", op1, op2, op3), "" };
+    return { 0, std::format("?TODO({:02X},{:02X},{:02X})", op1, op2, op3), "" };
 }
 } // namespace
 
@@ -670,10 +670,10 @@ void Writer::dump(int start, size_t bytes) const
     {
         if ((offset % 16) == 0)
         {
-            m_os << fmt::format("{:04X}:", base + offset);
+            m_os << std::format("{:04X}:", base + offset);
         }
         const auto b = m_memory.read_byte(base + offset);
-        hex_bytes += fmt::format(" {:02X}", b);
+        hex_bytes += std::format(" {:02X}", b);
         ascii_bytes += std::string(1, ((b < 0x20) || (b > 0x7f)) ? '.' : b);
         if (((offset + 1) % 16) == 0)
         {
@@ -692,12 +692,12 @@ void Writer::dump(int start, size_t bytes) const
 
 void Writer::display(std::uint16_t address, std::string_view s1, std::string_view s2) const
 {
-    m_os << fmt::format("{:04X}     {:<5}{}", address, s1, s2) << std::endl;
+    m_os << std::format("{:04X}     {:<5}{}", address, s1, s2) << std::endl;
 }
 
 void Writer::display(const zcpm::Registers& registers, std::string_view s1, std::string_view s2, const std::uint16_t offset) const
 {
-    m_os << fmt::format("{} A={:02X} B={:04X} D={:04X} H={:04X} S={:04X} P={:04X}  {:<5}{}",
+    m_os << std::format("{} A={:02X} B={:04X} D={:04X} H={:04X} S={:04X} P={:04X}  {:<5}{}",
                         flags_to_string(registers.AF & 0xFF),
                         registers.AF >> 8,
                         registers.BC,
@@ -709,7 +709,7 @@ void Writer::display(const zcpm::Registers& registers, std::string_view s1, std:
                         s2);
 
     m_os << std::endl;
-    m_os << fmt::format("{} '={:02X} '={:04X} '={:04X} '={:04X} X={:04X} Y={:04X}",
+    m_os << std::format("{} '={:02X} '={:04X} '={:04X} '={:04X} X={:04X} Y={:04X}",
                         flags_to_string(registers.altAF & 0xFF),
                         registers.altAF >> 8,
                         registers.altBC,
